Move utilities to RecoDecay
