+++ Analysis/Tasks/vertexerhf.cxx
@@ -12,6 +12,7 @@
 #include "Framework/AnalysisDataModel.h"
 #include "Framework/ASoAHelpers.h"
 #include "Analysis/SecondaryVertex.h"
+#include "Analysis/RecoDecay.h"
 #include "DetectorsVertexing/DCAFitterN.h"
 #include "ReconstructionDataFormats/Track.h"
